DN-DETR/smrc/requirements.txt
lxml
numpy
opencv-contrib-python
tqdm
scipy
#pygame~=1.9.6
opencv-python
matplotlib
scikit-learn>=0.24.1  # modified 2021-8-15
imagesize
filetype
xlsxwriter
Pillow
ffmpeg
moviepy
PySimpleGUI
